Reformatted the project

+++ src/loss.cpp
@@ -10,7 +10,7 @@ double mse(const Vector & y_true, const Vector & y_pred)
     {
         sum += std::pow(y_true[i] - y_pred[i], 2);
     }
-    return sum / y_true.size();
+    return sum / static_cast<double>(y_true.size());
 }
 
 double mae(const Vector &y_true, const Vector &y_pred)
@@ -20,7 +20,7 @@ double mae(const Vector &y_true, const Vector &y_pred)
     {
         sum += std::abs(y_true[i] - y_pred[i]);
     }
-    return sum / y_true.size();
+    return sum / static_cast<double>(y_true.size());
 }
 
 double cross_entropy(const Vector &y_true, const Vector &y_pred)
@@ -30,7 +30,7 @@ double cross_entropy(const Vector &y_true, const Vector &y_pred)
     {
         sum += y_true[i] * std::log(y_pred[i]);
     }
-    return -sum / y_true.size();
+    return -sum / static_cast<double>(y_true.size());
 }
 
 // initialize the static variables
